+++ providers/src/main/kotlin/org/agrfesta/sh/api/providers/switchbot/devices/SwitchBotMiniHub.kt
@@ -8,5 +8,5 @@ class SwitchBotMiniHub(
     override val uuid: UUID,
     override val deviceProviderId: String
 ) : DeviceDriver {
-    override val provider: Provider = Provider.NETATMO
+    override val provider: Provider = Provider.SWITCHBOT
 }